+++ tools/gulp/packages/info-package.ts
@@ -1,5 +1,4 @@
 import Package from './package';
-import {log,colors} from 'gulp-util';
 
 import * as data from '../../../package.json';