+++ tools/gulp/tasks/test-task.ts
@@ -5,7 +5,7 @@ import { WebpackKarmaPackage } from '../packages/webpack-package';
 
 export function createKarmaWebpackTask(karmaPack: WebpackKarmaPackage): void {
   task(`${karmaPack.getName()}:test`, (cb: (err?: any) => void) => {
-    //Add webpack config to karma
+    // Add webpack config to karma
     const karmaConfig: any = Object.assign({}, {
       files: [{ pattern: './src/test.ts', watched: false }],
       // preprocess matching files before serving them to the browser
@@ -16,7 +16,7 @@ export function createKarmaWebpackTask(karmaPack: WebpackKarmaPackage): void {
       webpack: karmaPack.getConfig(),
       configFile: karmaPack.resolveInProject('karma.config.js')
     });
-    //Start karma server
+    // Start karma server
     new Server(karmaConfig, cb).start();
   });
 }