+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+ - add tslint configuration file
+ - add protractor e2e configuration
  - add first component test
  - add test karma+jasmine configuration
  - create initial configuration using the bash script angular_setup